fix(shared): 修复Likable字段类型为int64
将LikeCount和ReplyCount字段类型从int改为int64，以匹配：
- MongoDB返回的数值类型
- 测试期望的类型
- 更适合大数值场景

影响范围：
- models/shared/social.go
- 所有使用Likable嵌入的结构体

+++ models/shared/social.go
@@ -2,7 +2,7 @@ package shared
 
 // Likable 支持点赞的实体混入
 type Likable struct {
-	LikeCount int `bson:"like_count" json:"likeCount"`
+	LikeCount int64 `bson:"like_count" json:"likeCount"`
 }
 
 // AddLike 增加点赞计数
@@ -21,7 +21,7 @@ func (l *Likable) RemoveLike() {
 type ThreadedConversation struct {
 	ParentID   *string `bson:"parent_id,omitempty" json:"parentId,omitempty"`
 	RootID     *string `bson:"root_id,omitempty" json:"rootId,omitempty"`
-	ReplyCount int     `bson:"reply_count" json:"replyCount"`
+	ReplyCount int64   `bson:"reply_count" json:"replyCount"`
 }
 
 // AddReply 增加回复计数
